pyproject.toml
[project]
name = "fictive"
version = "0.1.0"
description = "An interactive fiction tool based on state machines"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mypy>=1.17.1",
    "pyyaml>=6.0.2",
    "textual-dev>=1.7.0",
]
